pyproject.toml
[project]
name = "pocketdvdnet"
version = "0.1.0"
description = "Paper-only PocketDVDNet reproduction workflow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.8",
    "natsort>=8.4.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
    "scikit-image>=0.19.0",
    "scipy>=1.7.0",
    "tensorboardx>=2.5.0",
    "torch>=1.13.0",
    "torchaudio>=0.13.0",
    "torchvision>=0.14.0",
    "tqdm>=4.62.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "artifacts",
    "checkpoints",
    "external",
    "results",
]

[tool.ruff.lint]
ignore = ["F403", "F405"]
